fix warning and change real_type to double

+++ types.h
@@ -4,7 +4,7 @@ namespace treescript
 {
 
 typedef long long int int_type;
-typedef long double real_type;
+typedef double real_type;
 typedef bool bool_type;
 
 }
